generalize ShallowCopyIterator to UpdatingIterator
