+++ src/main/java/org/openhds/repository/results/ShallowCopyIterator.java
@@ -3,59 +3,19 @@ package org.openhds.repository.results;
 import org.openhds.domain.contract.UuidIdentifiable;
 import org.openhds.domain.util.ShallowCopier;
 
-import java.util.ArrayList;
-import java.util.Iterator;
-import java.util.List;
-
 /**
  * Created by Ben on 6/23/15.
  *
  * Wrap an existing EntityIterator and convert the results to shallow copies.
  *
  */
-public class ShallowCopyIterator<T extends UuidIdentifiable> implements EntityIterator<T>, Iterator<T> {
-
-    private final EntityIterator<T> original;
+public class ShallowCopyIterator<T extends UuidIdentifiable> extends UpdatingIterator<T> {
 
     public ShallowCopyIterator(EntityIterator<T> original) {
-        this.original = original;
+        super(original, ShallowCopyIterator::shallowCopy);
     }
 
-    @Override
-    public List<T> toList() {
-        List<T> list = new ArrayList<>();
-        for (T entity : original) {
-            list.add(shallowCopy(entity));
-        }
-        return list;
-    }
-
-    @Override
-    public String getCollectionName() {
-        return original.getCollectionName();
-    }
-
-    @Override
-    public void setCollectionName(String collectionName) {
-        original.setCollectionName(collectionName);
-    }
-
-    @Override
-    public Iterator<T> iterator() {
-        return this;
-    }
-
-    @Override
-    public boolean hasNext() {
-        return original.iterator().hasNext();
-    }
-
-    @Override
-    public T next() {
-        return shallowCopy(original.iterator().next());
-    }
-
-    private T shallowCopy(T entity) {
+    private static <T extends UuidIdentifiable> T shallowCopy(T entity) {
         return ShallowCopier.makeShallowCopy(entity, null);
     }
 }
